test(vhd-lib): from Jest to test
