+++ packages/vhd-lib/package.json
@@ -37,10 +37,12 @@
     "execa": "^5.0.0",
     "get-stream": "^6.0.0",
     "rimraf": "^5.0.1",
+    "test": "^3.3.0",
     "tmp": "^0.2.1"
   },
   "scripts": {
-    "postversion": "npm publish"
+    "postversion": "npm publish",
+    "test-integration": "node--test ./*.integ.js tests/*.integ.js Vhd/*.integ.js"
   },
   "author": {
     "name": "Vates SAS",
